Use Mutex instead of LockedState
