+++ Sources/Subprocess/Platforms/Subprocess+Unix.swift
@@ -433,38 +433,6 @@ extension FileDescriptor {
         }
     }
 
-    internal func readUntilEOF(upToLength maxLength: Int) async throws -> Data {
-        return try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<Data, any Error>) in
-            let dispatchIO = DispatchIO(
-                type: .stream,
-                fileDescriptor: self.rawValue,
-                queue: .global()
-            ) { error in
-                if error != 0 {
-                    continuation.resume(throwing: POSIXError(.init(rawValue: error) ?? .ENODEV))
-                }
-            }
-            var buffer: Data = Data()
-            dispatchIO.read(
-                offset: 0,
-                length: maxLength,
-                queue: .global()
-            ) { done, data, error in
-                guard error == 0 else {
-                    continuation.resume(throwing: POSIXError(.init(rawValue: error) ?? .ENODEV))
-                    return
-                }
-                if let data = data {
-                    buffer += Data(data)
-                }
-                if done {
-                    dispatchIO.close()
-                    continuation.resume(returning: buffer)
-                }
-            }
-        }
-    }
-
     internal func write<WriteSequence: Sequence & Sendable>(
         _ data: WriteSequence
     ) async throws where WriteSequence.Element == UInt8 {
